requirements.txt
numpy
matplotlib
openai
python-dotenv